+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2.1.1
+
+### Changes
+
+- `Banner` changed to have optional text.
+
+To make room for arbitrary children under header if no text provided.
+
+- `Banner` children (i.e bottom content) is now in proper layout under header and text.
+- Added story for Banner with both header, text and bottom content.
+
 ## 2.1.0
 
 ### Breaking changes